Remove casting
