+++ apps/hubspot/src/utils/utils.ts
@@ -1,5 +1,3 @@
-import { FieldType } from '@contentful/app-sdk';
-
 export const HUBSPOT_PRIVATE_APPS_URL = 'https://developers.hubspot.com/docs/api/private-apps';
 
 export const CONFIG_SCREEN_INSTRUCTIONS = [
@@ -20,18 +18,3 @@ export type AppInstallationParameters = {
 };
 
 export const CONFIG_CONTENT_TYPE_ID = 'hubspotConfig';
-
-export type SdkField = {
-  type: FieldType;
-  id: string;
-  uniqueId: string;
-  name: string;
-  locale?: string;
-  linkType?: string; // FieldLinkType
-  items?: {
-    type: string;
-    linkType: string;
-  }; // Items
-  supported: boolean;
-  value: any;
-};
